claude-code-agents-python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-code-agents-python"
version = "0.1.0"
description = "Agent SDK orchestration for iterative crawl campaigns, DSPy classification, and multi-language plugin generation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Claude Code Agents Team"},
]
dependencies = [
    "anthropic>=0.40",
    "dspy-ai>=2.5",
    "pydantic>=2.0",
    "click>=8.0",
    "scrapy>=2.11",
    "pyyaml",
    "jinja2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
claude-agents = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
